Don't suspend threads when injecting
cmd.exe is a singlethreaded application. When it is waiting for another
process (such as clink) to finish, it is blocked in
WaitForSingleObject().

That means there's no need to suspend its threads before injecting.

This is useful as enumerating all the threads on the machine frequently
takes around a third of a second, and this cost is incurred even if cmd
is started with /c, meaning it slows down script execution in
non-interactive shells as well.